Database spec is passing!
I used Jasmine to make the spec. There was one problem left: the JSON
was imported as string and not parsed. Plus it had comments in it, which
is a no-go for JSON. Now however, it works fine. I really like behaviour
driven development! I now know that my Database module functions fine.
And with *Jasmine* with **auto-update** enabled, I can **immediately**
see the results of my code updates!