Sliders working.
The sliders are working perfectly, their data coming from the part
metadata in the database. I even implemented a loading bar, which works
great, but isn't even visible, probably only on big models. On
slideStop, the hash is changed and the model is reloaded. I'm changing
the hash and not using pushState() because the latter doesn't allow the
user to copypaste the URL and send it to a friend. The next step is to
implement the parameters in the backend.

Added icons to illustrate how to control the CADview.
Added parameters and presets to db metadata object.
Added bootstrap.js. (For alert dismissal.)
Updated slider.js with seirya-bootstrap-slider, it's better than the
original, less buggy.